refactor(cells): move getCellColors + thresholds helpers into cells/ (refs #298)
